Rename integration package to tests
