+++ crates/flashblocks-rpc/src/lib.rs
@@ -2,6 +2,5 @@ pub mod cache;
 pub mod flashblocks;
 mod metrics;
 pub mod rpc;
-
 #[cfg(test)]
-mod integration;
+mod tests;
