+++ crates/flashblocks-rpc/src/tests/assets/genesis.json
